# lib/api-zod/src/generated/types/createContactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for Atendimento Compartilhado, a unified WhatsApp Business support inbox.
 * OpenAPI spec version: 0.2.0
 */

export interface CreateContactInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 8 */
  phoneNumber: string;
  /** @nullable */
  numberId?: string | null;
}