# lib/api-zod/src/generated/types/createUserInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for Atendimento Compartilhado, a unified WhatsApp Business support inbox.
 * OpenAPI spec version: 0.2.0
 */
import type { CreateUserInputRole } from './createUserInputRole';

export interface CreateUserInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 4 */
  password: string;
  role: CreateUserInputRole;
  numberIds?: string[];
}